+++ test/type/type_union/usable_as_test.exs
@@ -3,7 +3,7 @@ defmodule TypeTest.TypeUnion.UsableAsTest do
 
   @moduletag :usable_as
 
-  import Type, only: [builtin: 1]
+  import Type, only: :macros
   use Type.Operators
 
   describe "for union types" do